# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Filmfy",
        "version": "1.0"
    },
    "paths": {
        "/api/create-list": {
            "post": {
                "tags": [
                    "Lists"
                ],
                "summary": "Create a list",
                "operationId": "d53183bf7c45fa81c78e716962e02865",
                "requestBody": {
                    "description": "Pass list data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "users_id",
                                    "title",
                                    "description",
                                    "movies"
                                ],
                                "properties": {
                                    "users_id": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "movies": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfuly added"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Mostrar peliculas",
                "operationId": "27305720d93607b559f6a9b5d7ea64e4",
                "responses": {
                    "200": {
                        "description": "Mostrar todas las peliculas."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Mostrar una película por ID",
                "operationId": "3a532e24d1f50aef9ab58ca1e5074f31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    }
}